Добавляет отрисовку маршрута в отдельный presenter
